pyproject.toml
[tool.poetry]
name = "productdate"
version = "0.1.0"
description = ""
authors = ["kuptsovsasha <kuptsovsasha@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.2"
django-environ = "^0.9.0"
psycopg2-binary = "^2.9.4"
isort = "^5.10.1"
flake8 = "^5.0.4"
black = "^22.10.0"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.4"
django-filter = "^22.1"
django-rest-passwordreset = "^1.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[flake8]
ignore = '''
    # To be PEP-8 compliant, Black splits lines before binary operators
    W503,
    # For the same reason, Black enforces whitespace around slice operators
    E203,
    # Do not enforce newlines before class docstrings
    D203
'''
exclude = '''
    .git
    __pycache__,
    dist,
    build,
    provision,
    venv,
    settings.py,
    migrations
'''
# Black defaults to 88 and has some pretty good reasons for it
# a.k.a: Raymond Hettinger says so:
# https://www.youtube.com/watch?v=wf-BqAjZb8M&feature=youtu.be&t=260
max-line-length=88


[tool.black]
exclude = '''
    (
        \.eggs|
        \.git|
        \.hg|
        \.mypy_cache|
        \.nox|
        \.tox|
        \.venv|
        \.svn|
        _build|
        buck-out|
        build|
        dist|
        settings\.py|
        migrations|
        snapshots
    )/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap= 0
use_parentheses = true
line_length = 88
skip_glob = '*migrations*,*snapshot*'